+++ core/src/elements/ons-tabbar/index.js
@@ -378,7 +378,14 @@ export default class TabbarElement extends BaseElement {
           p = 1; // Visual fix for some devices
         }
 
-        this.style.top = top ? parseInt(window.getComputedStyle(page._getContentElement(), null).getPropertyValue('padding-top'), 10) - p + 'px' : '';
+        const content = page._getContentElement();
+        const cs = window.getComputedStyle(page._getContentElement(), null)
+
+        this.style.top = top ? parseInt(cs.getPropertyValue('padding-top'), 10) - p + 'px' : '';
+
+        // Refresh content top - Fix for iOS 8
+        content.style.top = cs.top;
+        content.style.top = '';
       });
     }